requirements.txt
pandas==1.3.4
Pillow==9.0.1
torch==1.10.2+cu113
torchvision==0.11.3+cu113
